requirements.txt
Flask
SQLAlchemy
python-dotenv 
pymysql
alembic
Flask-Migrate
Flask-Bcrypt
PyJWT
brazilcep
flasgger